Fronend can now emit Quantum IR, too
