+++ frontend/qasm/qasmtoqillr.py
@@ -1,0 +1,470 @@
+"""
+#   Frontend generating QILLR dialect code from QASM2 and QASM3 code.
+#
+# @author  Washim Neupane (washim.neupane@outlook.com)
+# @author  Lars Schütze (lars.schuetze@tu-dresden.de)
+"""
+
+from __future__ import annotations
+
+import re
+from enum import Enum
+from functools import reduce
+
+from mlir._mlir_libs._mlirDialectsQILLR import QubitType
+from mlir._mlir_libs._mlirDialectsQILLR import qillr as qillrdialect
+from mlir.dialects import arith, func, qillr, scf, tensor
+from mlir.dialects.builtin import Block, IntegerType, RankedTensorType
+from mlir.ir import Context, F64Type, InsertionPoint, Location, Module, StringAttr, TypeAttr, Value
+from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
+from qiskit.circuit import Clbit, Instruction, Operation, ParameterExpression, Qubit
+from qiskit.circuit import library as lib
+from qiskit.circuit.classical.expr import Expr
+from qiskit.circuit.controlflow.if_else import IfElseOp
+from qiskit.qasm2 import loads as qasm2_loads
+from qiskit.qasm2.parse import LEGACY_CUSTOM_INSTRUCTIONS
+from qiskit.qasm2.parse import _DefinedGate as QASM2_Gate
+
+
+class ConversionError(RuntimeError): ...
+
+
+class ParseError(RuntimeError): ...
+
+
+# Specify the QASM version the frontend conforms to
+class QASMVersion(Enum):
+    Unspecified = 0
+    QASM_2_0 = 1
+    QASM_3_0 = 2
+
+
+# Types that can be coerced to a valid Qubit specifier in a circuit.
+type QubitSpecifier = Qubit | QuantumRegister
+# int,
+# slice,
+# Sequence[Union[Qubit, int]],
+
+# Types that can be coerced to a valid Clbit specifier in a circuit.
+type ClbitSpecifier = Clbit | ClassicalRegister
+# int,
+# slice,
+# Sequence[Union[Clbit, int]],
+
+
+class Scope:
+    def __init__(self, visited: dict[str, qillr.GateOp] | None = None) -> None:
+        self.qregs: dict[str, qillr.AllocOp] = {}
+        self.cregs: dict[str, qillr.AllocResultOp] = {}
+        self.reads: dict[str, qillr.ReadMeasurementOp] = {}
+        self.visitedGates: dict[str, qillr.GateOp] = visited if visited is not None else {}
+
+    @classmethod
+    def fromList(
+        cls,
+        qregs: list[QubitSpecifier],
+        cregs: list[ClbitSpecifier],
+        visited: dict[str, qillr.GateOp] | None = None,
+        reads: dict[str, qillr.ReadMeasurementOp] | None = None,
+    ) -> Scope:
+        s = cls(visited)
+        s.qregs = {str(q): None for qreg in qregs for q in qreg}
+        s.cregs = {str(c): None for creg in cregs for c in creg}
+        s.reads = {str(r): None for read in reads for r in read} if reads is not None else {}
+        return s
+
+    @classmethod
+    def fromMap(
+        cls,
+        qregs: dict[str, qillr.AllocOp],
+        cregs: dict[str, qillr.AllocResultOp],
+        visited: dict[str, qillr.GateOp] | None = None,
+        reads: dict[str, qillr.ReadMeasurementOp] | None = None,
+    ) -> Scope:
+        s = cls(visited)
+        s.qregs = qregs
+        s.cregs = cregs
+        s.reads = reads if reads is not None else {}
+        return s
+
+    def findAlloc(self, reg: QubitSpecifier) -> qillr.AllocOp:
+        return self.qregs.get(str(reg))
+
+    def setAlloc(self, reg: QubitSpecifier, alloc: qillr.AllocOp) -> None:
+        self.qregs[str(reg)] = alloc
+
+    def findResult(self, reg: ClbitSpecifier) -> qillr.AllocResultOp:
+        return self.cregs.get(str(reg))
+
+    def setResult(self, reg: ClbitSpecifier, ralloc: qillr.AllocResultOp) -> None:
+        self.cregs[str(reg)] = ralloc
+
+    def findGate(self, gate: QASM2_Gate) -> qillr.GateOp:
+        return self.visitedGates.get(str(gate.name))
+
+    def setGate(self, gate: QASM2_Gate, newGate: qillr.GateOp) -> None:
+        self.visitedGates[str(gate.name)] = newGate
+
+    def findRead(self, reg: ClbitSpecifier) -> qillr.ReadMeasurementOp:
+        return self.reads.get(str(reg))
+
+    def setRead(self, reg: ClbitSpecifier, read: qillr.ReadMeasurementOp) -> None:
+        self.reads[str(reg)] = read
+
+
+class QASMToMLIRVisitor:
+    def __init__(self, compat: QASMVersion, context: Context, module: Module, loc: Location, block: Block, scope: Scope) -> None:
+        self.compat: QASMVersion = compat
+        self.context: Context = context
+        self.module: Module = module
+        self.loc: Location = loc
+        self.block: Block = block
+        self.scope = scope
+        self.resetQubits: dict[Value, bool] = {}
+
+    @classmethod
+    def fromParent(cls, parent: QASMToMLIRVisitor, *, block: Block | None = None, scope: Scope | None = None):
+        return cls(
+            parent.compat,
+            parent.context,
+            parent.module,
+            parent.loc,
+            parent.block if block is None else block,
+            parent.scope if scope is None else scope,
+        )
+
+    def visitCircuit(self, circuit: QuantumCircuit) -> None:
+        for instr in circuit.data:
+            if isinstance(instr.operation, Expr):
+                self.visitClassic(instr)
+            elif isinstance(instr.operation, Instruction):
+                self.visitInstruction(instr.operation, instr.qubits, instr.clbits)
+            elif isinstance(instr.operation, Operation):
+                self.visitQuantum(instr.operation)
+            else:
+                raise ParseError(f"Unknown instruction: {instr} of type {type(instr)}")
+
+    def visitQuantumBit(self, reg: Qubit, *, emitReset=False) -> Value:
+        if self.scope.findAlloc(reg) is None:
+            alloc: qillr.AllocOp = qillr.AllocOp(loc=self.loc, ip=InsertionPoint(self.block))
+            self.resetQubits[alloc.result] = True
+            self.scope.setAlloc(reg, alloc.result)
+
+        allocResult: qillr.AllocOp = self.scope.findAlloc(reg)
+        if emitReset and self.resetQubits[allocResult]:
+            self.resetQubits[allocResult] = False
+            qillr.ResetOp(allocResult, loc=self.loc, ip=InsertionPoint(self.block))
+        return allocResult
+
+    def visitClassicalBit(self, reg: Clbit) -> Value:
+        if self.scope.findResult(reg) is None:
+            ralloc: qillr.AllocResultOp = qillr.AllocResultOp(loc=self.loc, ip=InsertionPoint(self.block))
+            self.scope.setResult(reg, ralloc.result)
+            return ralloc.result
+
+        return self.scope.findResult(reg)
+
+    def visitClassic(self, expr: Expr) -> Value:
+        if isinstance(expr, ParameterExpression):
+            raise NotImplementedError("Parameter Expression")
+        elif isinstance(expr, float):
+            return arith.ConstantOp(F64Type.get(self.context), expr, ip=InsertionPoint(self.block)).result
+        else:
+            raise NotImplementedError(f"Classic expressions are not supported for {expr}")
+
+    # Operation encapsulates virtual instructions that must
+    # be synthesized to physical instructions
+    def visitQuantum(self, instr: Operation) -> None:
+        raise NotImplementedError(f"Virtual quantum expressions are not supported for {instr}")
+
+    # Instruction represents physical quantum instructions
+    def visitInstruction(self, instr: Instruction, qubits: list[QubitSpecifier], clbits: list[ClbitSpecifier]) -> None:
+        match instr, len(qubits):
+            case lib.Barrier(), _:
+                with self.loc:
+                    args = [self.visitQuantumBit(q) for q in qubits]
+                    qillr.BarrierOp(args, ip=InsertionPoint(self.block))
+            case QASM2_Gate(), _:
+                self._visitDefinedGate(instr, qubits, clbits)
+            case IfElseOp(), _:
+                self._visitIfElse(instr, qubits, clbits)
+            case lib.CCXGate(), 3:
+                with self.loc:
+                    control1: Value = self.visitQuantumBit(qubits[0])
+                    control2: Value = self.visitQuantumBit(qubits[1])
+                    target: Value = self.visitQuantumBit(qubits[2])
+                    qillr.CCXOp(control1, control2, target, ip=InsertionPoint(self.block))
+            case lib.CSwapGate(), 3:
+                with self.loc:
+                    control: Value = self.visitQuantumBit(qubits[0])
+                    lhs: Value = self.visitQuantumBit(qubits[1])
+                    rhs: Value = self.visitQuantumBit(qubits[2])
+                    qillr.CSwapOp(control, lhs, rhs, ip=InsertionPoint(self.block))
+            case _, 1:
+                self._visitUnaryGates(instr, qubits, clbits)
+            case _, 2:
+                self._visitBinaryGates(instr, qubits, clbits)
+            case _, _:
+                raise NotImplementedError(f"{instr} of type {type(instr)}")
+
+    def _visitUnaryGates(self, instr: Instruction, qubits: list[QubitSpecifier], clbits: list[ClbitSpecifier]) -> None:
+        assert len(qubits) == 1, f"Require unary gate, got: {instr}"
+        match qubits[0]:
+            case QuantumRegister():
+                raise NotImplementedError(f"Visiting instruction {instr} with QuantumRegister {qubits} and {clbits}")
+            case Qubit():
+                with self.loc:
+                    target: Value = self.visitQuantumBit(qubits[0])
+                    match instr:
+                        case lib.XGate():
+                            qillr.XOp(target, ip=InsertionPoint(self.block))
+                        case lib.YGate():
+                            qillr.YOp(target, ip=InsertionPoint(self.block))
+                        case lib.ZGate():
+                            qillr.ZOp(target, ip=InsertionPoint(self.block))
+                        case lib.HGate():
+                            qillr.HOp(target, ip=InsertionPoint(self.block))
+                        case lib.SGate():
+                            qillr.SOp(target, ip=InsertionPoint(self.block))
+                        case lib.SXGate():
+                            qillr.SXOp(target, ip=InsertionPoint(self.block))
+                        case lib.SdgGate():
+                            qillr.SdgOp(target, ip=InsertionPoint(self.block))
+                        case lib.TGate():
+                            qillr.TOp(target, ip=InsertionPoint(self.block))
+                        case lib.TdgGate():
+                            qillr.TdgOp(target, ip=InsertionPoint(self.block))
+                        case lib.RZGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.RzOp(target, angle, ip=InsertionPoint(self.block))
+                        case lib.RXGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.RxOp(target, angle, ip=InsertionPoint(self.block))
+                        case lib.RYGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.RyOp(target, angle, ip=InsertionPoint(self.block))
+                        case lib.U3Gate():
+                            theta, phi, lam = [self.visitClassic(param) for param in instr.params]
+                            qillr.U3Op(target, theta, phi, lam, ip=InsertionPoint(self.block))
+                        case lib.U2Gate():
+                            phi, lam = [self.visitClassic(param) for param in instr.params]
+                            qillr.U2Op(target, phi, lam, ip=InsertionPoint(self.block))
+                        case lib.U1Gate():
+                            lam = self.visitClassic(instr.params[0])
+                            qillr.U1Op(target, lam, ip=InsertionPoint(self.block))
+                        case lib.UGate():
+                            # TODO: https://docs.quantum.ibm.com/api/qiskit/qiskit.circuit.library.UGate
+                            theta, phi, lam = [self.visitClassic(param) for param in instr.params]
+                            qillr.U3Op(target, theta, phi, lam, ip=InsertionPoint(self.block))
+                        case lib.Reset():
+                            self._visitReset(target)
+                        case lib.Measure():
+                            bit: Value = self.visitClassicalBit(clbits[0])
+                            qillr.MeasureOp(target, bit, ip=InsertionPoint(self.block))
+                            self._emitReadOrVal(clbits[0], override=True)
+                        case lib.IGate():
+                            qillr.IdOp(target, ip=InsertionPoint(self.block))
+                        case lib.PhaseGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.PhaseOp(target, angle, ip=InsertionPoint(self.block))
+                        case _:
+                            raise NotImplementedError(f"Unary gate {instr}")
+
+    def _visitBinaryGates(self, instr: Instruction, qubits: list[QubitSpecifier], clbits: list[ClbitSpecifier]) -> None:
+        assert len(qubits) == 2, f"Require binary gate, got: {instr}"
+        match qubits[0], qubits[1]:
+            case QuantumRegister(), QuantumRegister():
+                raise NotImplementedError(f"Visiting instruction {instr} with QuantumRegister {qubits} and {clbits}")
+            case Qubit(), Qubit():
+                with self.loc:
+                    lhs: Value = self.visitQuantumBit(qubits[0])
+                    rhs: Value = self.visitQuantumBit(qubits[1])
+                    match instr:
+                        case lib.SwapGate():
+                            qillr.SwapOp(lhs, rhs, ip=InsertionPoint(self.block))
+                        case lib.CZGate():
+                            qillr.CZOp(lhs, rhs, ip=InsertionPoint(self.block))
+                        case lib.CXGate():
+                            qillr.CNOTOp(lhs, rhs, ip=InsertionPoint(self.block))
+                        case lib.CRYGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.CRyOp(lhs, rhs, angle, ip=InsertionPoint(self.block))
+                        case lib.CRZGate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.CRzOp(lhs, rhs, angle, ip=InsertionPoint(self.block))
+                        case lib.CU1Gate():
+                            angle = self.visitClassic(instr.params[0])
+                            qillr.CU1Op(lhs, rhs, angle, ip=InsertionPoint(self.block))
+
+    def _visitDefinedGate(self, instr: QASM2_Gate, qubits: list[QubitSpecifier], clbits: list[ClbitSpecifier]) -> None:
+        if instr.definition is not None:
+            if self.scope.findGate(instr) is None:
+                # Construct qillr.GateOp for defined custom gate
+                # Insert into module body and recursively visit gate body
+                inputs: list[QubitType] = [QubitType.get(self.context) for _ in range(instr.num_qubits)]
+                gty: func.FunctionType = func.FunctionType.get(inputs=inputs, results=[], context=self.context)
+                gate: qillr.GateOp = qillr.GateOp(
+                    StringAttr.get(str(instr.name)),
+                    TypeAttr.get(gty),
+                    loc=self.loc,
+                    ip=InsertionPoint.at_block_begin(self.module.body),
+                )
+                arg_locs = [self.loc for _ in inputs]
+                gate.body.blocks.append(*inputs, arg_locs=arg_locs)
+                gateBody: Block = gate.body.blocks[0]
+
+                self.scope.setGate(instr, gate)
+                circuit: QuantumCircuit = instr.definition
+                gateQregs = {str(q): a for q, a in zip(circuit.qubits, gate.body.blocks[0].arguments)}
+                innerGateScope: Scope = Scope.fromMap(gateQregs, {}, self.scope.visitedGates)
+                visitor: QASMToMLIRVisitor = QASMToMLIRVisitor.fromParent(self, block=gateBody, scope=innerGateScope)
+                visitor.visitCircuit(circuit)
+                qillr.ReturnOp(loc=self.loc, ip=InsertionPoint(gateBody))
+            # Construct qillr.CallOp for defined custom gate
+            callee: StringAttr = instr.name
+            operands: list[Value] = [self.visitQuantumBit(q) for q in qubits]
+            qillr.GateCallOp(callee, operands, loc=self.loc, ip=InsertionPoint(self.block))
+        else:
+            ParseError(f"Expected gate with definition, got: {instr}")
+
+    def _visitReset(self, qubit: Value) -> None:
+        self.resetQubits[qubit] = False
+        qillr.ResetOp(qubit, ip=InsertionPoint(self.block))
+
+    def _visitIfElse(self, instr: IfElseOp, qubits: list[QubitSpecifier], clbits: list[ClbitSpecifier]) -> None:
+        with self.loc:
+            condition: Value = self._visitIfElseCondition(instr.condition, qubits, clbits)
+            true_body, false_body = instr.params
+            hasElse: bool = false_body is not None
+
+            ifOp: scf.IfOp = scf.IfOp(condition, hasElse=hasElse, ip=InsertionPoint(self.block))
+            thenVisitor = QASMToMLIRVisitor.fromParent(self, block=ifOp.then_block)
+            thenVisitor.visitCircuit(true_body)
+            scf.YieldOp(results_=[], ip=InsertionPoint(ifOp.then_block))
+            if hasElse:
+                elseVisitor = QASMToMLIRVisitor.fromParent(self, block=ifOp.else_block)
+                elseVisitor.visitCircuit(false_body)
+                scf.YieldOp(results_=[], ip=InsertionPoint(ifOp.else_block))
+
+    def _emitReadOrVal(self, reg: Clbit, *, override: bool = False) -> Value:
+        if self.scope.findRead(reg) is None or override:
+            clval: Value = self.visitClassicalBit(reg)
+            read: qillr.ReadMeasurementOp = qillr.ReadMeasurementOp(clval, loc=self.loc, ip=InsertionPoint(self.block))
+            self.scope.setRead(reg, read.result)
+            return read.result
+
+        return self.scope.findRead(reg)
+
+    def _visitIfElseCondition(
+        self,
+        condition: Expr | tuple[ClassicalRegister, int] | tuple[Clbit, int],
+        qubits: list[QubitSpecifier],
+        clbits: list[ClbitSpecifier],
+    ) -> Value:
+        match condition:
+            case Expr():
+                raise NotImplementedError(f" IfElseOp with condition of type {type(condition)}")
+            case (bitOrRegister, axiom):  # tuple[ClassicalRegister, int] | tuple[Clbit, int]
+                with self.loc:
+                    i1Type = IntegerType.get_signless(1)
+                    match bitOrRegister:
+                        case Clbit():
+                            measurement: Value = self._emitReadOrVal(bitOrRegister)
+                            axiomval: Value = arith.ConstantOp(i1Type, axiom, ip=InsertionPoint(self.block)).result
+                            return arith.CmpIOp(
+                                arith.CmpIPredicate.eq, measurement, axiomval, ip=InsertionPoint(self.block)
+                            ).result
+                        case ClassicalRegister():
+                            clvals: list[tuple[int, Value]] = [
+                                (bit_at(axiom, i), clbit) for i, clbit in enumerate(clbits[0]._register)
+                            ]
+                            cmpis: list[Value] = []
+                            for b, clbit in clvals:
+                                measurement: Value = (
+                                    self._emitReadOrVal(clbit)
+                                    if self.scope.findRead(clbit) is not None
+                                    else arith.ConstantOp(i1Type, 0, ip=InsertionPoint(self.block)).result
+                                )
+                                axiomval: Value = arith.ConstantOp(i1Type, b, ip=InsertionPoint(self.block)).result
+                                cmpis.append(
+                                    arith.CmpIOp(
+                                        arith.CmpIPredicate.eq, measurement, axiomval, ip=InsertionPoint(self.block)
+                                    ).result
+                                )
+                            return reduce(
+                                lambda cmps, cmp: arith.AndIOp(cmps, cmp, ip=InsertionPoint(self.block)).result,
+                                cmpis[1:],  # rest of the list
+                                cmpis[0],  # initial value
+                            )
+                        case _:
+                            raise NotImplementedError(f"condition of type {type(condition)}")
+
+
+def qasm_version(code: str) -> QASMVersion:
+    match = re.search(r"OPENQASM\s+(\d+)\.(\d+);", code)
+
+    if match:
+        major = int(match.group(1))
+        minor = int(match.group(2))
+
+        if major == 2 and minor == 0:
+            return QASMVersion.QASM_2_0
+        elif major == 3 and minor == 0:
+            return QASMVersion.QASM_3_0
+
+    return QASMVersion.Unspecified
+
+
+def QASMToMLIR(code: str, emitResults: bool) -> Module:
+    compat: QASMVersion = qasm_version(code)
+    circuit: QuantumCircuit
+
+    match compat:
+        case QASMVersion.QASM_2_0:
+            try:
+                circuit = qasm2_loads(code, custom_instructions=LEGACY_CUSTOM_INSTRUCTIONS)
+            except Exception as e:
+                raise ConversionError(f"QASM2 parse failed: {e}")
+        case QASMVersion.QASM_3_0:
+            raise NotImplementedError("QASM3 not implemented")
+            # circuit = qasm3_loads(code)
+        case QASMVersion.Unspecified:
+            raise ParseError("No version string found")
+
+    context: Context = Context()
+    context.allow_unregistered_dialects = True
+    qillrdialect.register_dialect(context)
+
+    with context:
+        location: Location = Location.unknown()
+        module: Module = Module.create(location)
+
+        scope: Scope = Scope.fromList(circuit.qregs, circuit.cregs)
+
+        qasm_main: func.FuncOp = func.FuncOp("qasm_main", ([], []), visibility="public", loc=location)
+        qasm_main.add_entry_block()
+
+        visitor: QASMToMLIRVisitor = QASMToMLIRVisitor(compat, context, module, location, qasm_main.entry_block, scope)
+        visitor.visitCircuit(circuit)
+
+        for qubit in circuit.qubits:
+            visitor.visitQuantumBit(qubit, emitReset=True)
+
+        if not emitResults:
+            func.ReturnOp([], loc=location, ip=InsertionPoint(qasm_main.entry_block))
+        else:
+            # Create a new main function with the correct type and move the body to it
+            m: list[Value] = [r for _, r in scope.reads.items() if r is not None]
+            resType: RankedTensorType = RankedTensorType.get([len(m)], IntegerType.get_signless(1), loc=location)
+            qasm_main.attributes["function_type"] = TypeAttr.get(func.FunctionType.get([], [resType]))
+            # Merge all measurements into a tensor and return it
+            res: Value = tensor.FromElementsOp(resType, m, loc=location, ip=InsertionPoint(qasm_main.entry_block)).result
+            func.ReturnOp([res], loc=location, ip=InsertionPoint(qasm_main.entry_block))
+
+    module.body.append(qasm_main)
+    return module
+
+
+def bit_at(n: int, i: int) -> int:
+    """Return the value (0 or 1) of bit *i* of n."""
+    return (n >> i) & 1
