+++ lib/CAPI/Dialect/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(QILLR)
 add_subdirectory(RVSDG)
+add_subdirectory(Quantum)
