+++ lib/CAPI/Dialect/Quantum/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_mlir_upstream_c_api_library(MLIRCAPIQUANTUM
+  Quantum.cpp
+
+  LINK_LIBS PUBLIC
+    QuantumIR
+    MLIRCAPIIR
+)
