+++ python/mlir/dialects/QuantumOps.td
@@ -1,0 +1,12 @@
+//===- QuantumOps.td - Entry point for Quantum bindings -----*- tablegen -*-===//
+//                                                                               
+// @author  Lars Schütze (lars.schuetze@tu-dresden.de)                                 
+//===----------------------------------------------------------------------===// 
+
+
+#ifndef PYTHON_BINDINGS_QUANTUM_OPS
+#define PYTHON_BINDINGS_QUANTUM_OPS
+
+include "quantum-mlir/Dialect/Quantum/IR/QuantumOps.td"
+
+#endif // PYTHON_BINDINGS_QUANTUM_OPS
