+++ python/mlir/dialects/quantum.py
@@ -1,0 +1,6 @@
+#
+# ruff: noqa: F403
+# @author  Lars Schütze (lars.schuetze@tu-dresden.de)
+from .._mlir_libs._mlirDialectsQuantum import *  # pyright: ignore
+from .._mlir_libs._mlirDialectsQuantum.quantum import *  # pyright: ignore
+from ._quantum_ops_gen import *  # pyright: ignore
